Change new game message and add colour to text
Changed the refreshMessage text inside the playGame function that pops up when the game completes
so that it says to press New Game instead of refreshing
since a New Game button was added for users to refresh the page.
Also added colour to some of the text to make it more visually appealing.